+++ iai-callgrind/src/lib.rs
@@ -422,8 +422,9 @@ pub use iai_callgrind_macros::{binary_benchmark, library_benchmark};
 // documentation in `__internal::mod` for more details.
 #[cfg(feature = "default")]
 pub use iai_callgrind_runner::api::{
-    CachegrindMetric, CachegrindMetrics, CallgrindMetrics, DelayKind, DhatMetric, Direction,
-    EntryPoint, ErrorMetric, EventKind, FlamegraphKind, Pipe, Stdin, Stdio, ValgrindTool,
+    CachegrindMetric, CachegrindMetrics, CallgrindMetrics, DelayKind, DhatMetric, DhatMetrics,
+    Direction, EntryPoint, ErrorMetric, EventKind, FlamegraphKind, Limit, Pipe, Stdin, Stdio,
+    ValgrindTool,
 };
 #[cfg(feature = "default")]
 pub use lib_bench::LibraryBenchmarkConfig;